requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.5.0
matplotlib>=3.5.0
plotly>=5.0.0